Progressing with mongo support, random data supplier

+++ aggregator/converters/base.py
@@ -117,7 +117,36 @@ class BaseConverter(object):
         raise NotImplementedError('`dimensions` getter was not implemented')
 
     def data(self, v_name):
-        raise NotImplementedError('variable `data` getter was not implemented')
+        raise NotImplementedError('variable data getter was not implemented')
+
+    def count(self, v_name):
+        total = 1
+        dt = self.data(v_name)
+        while True:
+            try:
+                total *= len(dt)
+                dt = dt[0]
+            except:
+                break
+
+        return total
+
+    def get_value(self, v_name, comb):
+        value = self.data(v_name=v_name)
+        for idx, _ in enumerate(comb):
+            try:
+                value = value[comb[idx][0]]
+            except IndexError:
+                raise ValueError('Could not get value for this combination')
+
+        if str(value) == '--':
+            return None
+
+        return value
+
+    def normalize(self, dimension, value):
+        # default normalization changes nothing
+        return value
 
     def get_variable(self, v_name):
         try:
@@ -153,96 +182,6 @@ class BaseConverter(object):
 
         return os.path.join(dist_path, filename)
 
-    def predicted_document_size(self, dimension_sizes, shards):
-        # predicted size if no shards
-        full_document_size = reduce(lambda x, y: x*y, dimension_sizes) * self.AVERAGE_ELEMENT_SIZE
-
-        # divide based on shards
-        document_size = full_document_size
-        for shard in shards:
-            if shard:
-                document_size /= (len(shard) + 0.0)
-
-        return int(document_size)
-
-    def get_document_shards(self, dimension_sizes):
-        # predict document size
-        # initially plan for no sharding
-        shards = [[(0, dimension_sizes[idx])] for idx in xrange(dimension_sizes.__len__())]
-        shard_idx = 0
-
-        while self.predicted_document_size(dimension_sizes, shards) >= self.MAX_DOCUMENT_SIZE:
-
-            # check if dimension can even be sharded
-            if dimension_sizes[shard_idx] == 1 or \
-                    ((type(shards[shard_idx]) == list) and len(shards[shard_idx]) == dimension_sizes[shard_idx]):
-                shard_idx += 1
-                continue
-
-            # increase sharding on this dimension
-            n_of_shards = len(shards[shard_idx]) + 1
-            shard_size = dimension_sizes[shard_idx] / (n_of_shards + 0.0)
-            shards[shard_idx] = [(int(idx * shard_size), int((idx + 1) * shard_size))
-                                 for idx in xrange(n_of_shards)]
-
-        return shards
-
-    def json_documents(self, variable):
-        print variable.name
-        data = self.data(v_name=variable.name)
-
-        dim_lens = []
-        dt = data
-        while True:
-            if type(dt) not in [list, MaskedArray]:
-                break
-
-            dim_lens.append(dt.__len__())
-            dt = dt[0]
-
-        print dim_lens
-
-        shards = self.get_document_shards(dim_lens)
-
-        for shard in itertools.product(*shards):
-            # slices
-            slc = [slice(None)] * dim_lens.__len__()
-            for idx, s in enumerate(shard):
-                slc[idx] = slice(s[0], s[1])
-
-            encoded_document = encode_document({
-                'variable': variable.name,
-                'offsets': [s[0] for s in shard],
-                'values': data[slc]
-            })
-            yield encoded_document
-
-    def write_to_disk(self):
-
-        dist_path = self.full_output_path('.'.join(self.name.split('.')[:-1]))
-
-        # make sure the dataset folder exists
-        if not os.path.isdir(dist_path):
-            os.mkdir(dist_path)
-
-        with open('%s\\dataset_info.json' % dist_path, 'w') as output:
-            output.write(json.dumps(self.dataset.to_json()))
-
-        with open('%s\\variables.json' % dist_path, 'w') as output:
-            output.write(json.dumps([v.to_json() for v in self.variables]))
-
-        with open('%s\\dimensions.json' % dist_path, 'w') as output:
-            output.write(json.dumps([d.to_json() for d in self.dimensions]))
-
-        # make sure the actual data folder exists
-        if not os.path.isdir('%s\\data' % dist_path):
-            os.mkdir('%s\\data' % dist_path)
-
-        for variable in self.variables:
-            with open('%s\\data\\%s.json' % (dist_path, variable.name), 'w') as output:
-                for datum in self.json_documents(variable=variable):
-                    output.write(json.dumps(datum))
-
     def store(self, target, stdout=None):
         """
         :param target: Either {'type': 'postgres', 'cursor': <Cursor>, 'withIndices': True|False} or
@@ -257,7 +196,7 @@ class BaseConverter(object):
             elif target['type'] == 'mongo':
                 target['db'][v.name].insert_many(iv)
 
-        def db_serialize(val):
+        def postgres_serialize(val):
             if type(val) == datetime.datetime:
                 return "TIMESTAMP '%s'" % val.isoformat().replace('T', ' ')
             else:
@@ -288,7 +227,7 @@ class BaseConverter(object):
                                                     dataset=agd)
                 elif target['type'] == 'mongo':
                     v_doc = v.to_json()
-                    v_doc['dataset_id'] = agd['_id']
+                    v_doc['dataset_id'] = agd
                     agv = target['db'].variables.insert(v_doc)
 
                 dimensions = []
@@ -303,12 +242,13 @@ class BaseConverter(object):
                                                                    if d.step is not None else None,
                                                                    axis=d.axis,
                                                                    variable=agv)
+
                             elif target['type'] == 'mongo':
                                 d_doc = d.to_json()
-                                d_doc['variable_id'] = agv['_id']
+                                d_doc['variable_id'] = agv
                                 agdim = target['db'].dimensions.insert(d_doc)
 
-                            dimensions.append(agdim)
+                            dimensions.append(d)
                             break
 
                 # create data table for variable
@@ -316,58 +256,52 @@ class BaseConverter(object):
                     agv.create_data_table(cursor=target['cursor'], with_indices=target['with_indices'])
 
                 # add data
-                data = self.data(v_name=v.name)
                 dim_values = []
                 for dimension in dimensions:
                     vv = []
-                    v = dimension.min
+                    x = dimension.min
                     idx = 0
-                    while v <= dimension.max:
-                        vv.append((idx, self.normalize(dimension, v)))
+                    while x <= dimension.max:
+                        vv.append((idx, self.normalize(dimension, x)))
                         if dimension.step is None:
                             break
                         idx += 1
-                        v += dimension.step
+                        x += dimension.step
                     dim_values.append(vv)
 
                 insert_values = []
                 progress = 0
-                total = 1.0
-                dt = data
-                while True:
-                    try:
-                        total *= len(dt)
-                        dt = dt[0]
-                    except:
-                        break
+                total = self.count(v_name=v.name)
 
                 for comb in itertools.product(*dim_values):
-                    dt = data
                     error = False
-                    for idx, dimension in enumerate(comb):
-                        try:
-                            dt = dt[comb[idx][0]]
-                        except:
-                            error = True
-                            break
+                    value = None
+                    try:
+                        value = self.get_value(v_name=v.name, comb=comb)
+                    except ValueError:
+                        error = True
 
                     progress += 1
-                    if error:
+                    if error or (value is None):
                         continue
-                    if str(dt) != '--':
-                        if target['type'] == 'postgres':
-                            insert_values.append('(%s)' % ','.join([db_serialize(combi[1]) for combi in comb] + [str(dt)]))
-                        elif target['type'] == 'mongo':
-                            val_doc = {}
-                            for idx, combi in comb:
-                                val_doc[self.dimensions[idx].name] = combi[1]
 
-                            val_doc['value'] = dt
-                            insert_values.append(val_doc)
+                    if target['type'] == 'postgres':
+                        insert_values.append('(%s)' % ','.join([postgres_serialize(combi[1]) for combi in comb] +
+                                                               [str(value)]))
+                    elif target['type'] == 'mongo':
+                        val_doc = {}
+                        for idx, combi in enumerate(comb):
+                            val_doc[dimensions[idx].name] = combi[1]
+
+                        val_doc['value'] = value
+                        insert_values.append(encode_document(val_doc))
+
                     if len(insert_values) == 1000:
                         if stdout:
                             stdout.write("\r Adding data... %d%%" % (progress * 100 / total), ending='')
                             stdout.flush()
+                        else:
+                            print('%d%%' % (progress * 100 / total))
 
                         insert(insert_values)
                         insert_values = []
@@ -379,8 +313,10 @@ class BaseConverter(object):
                 if stdout:
                     stdout.write("\r Completed\t\t\t", ending='\n')
                     stdout.flush()
+                else:
+                    print('completed')
         except:
-            if agd:
+            if agd and type(agd) == AgDataset:
                 agd.delete()
 
             raise
